+++ editor/templates/index.js
@@ -1,4 +1,0 @@
-import Empty from './empty';
-import Example from './example';
-
-export default [Empty, Example];