+++ packages/app/components/TitleBar/TitleBar.jsx
@@ -1,0 +1,42 @@
+import { NavbarBrand, NavbarItem } from '@appsemble/react-bulma';
+import PropTypes from 'prop-types';
+import React from 'react';
+
+import Portal from '../Portal';
+import SideMenuButton from '../SideMenuButton';
+import styles from './TitleBar.css';
+
+/**
+ * The title bar on the top of the page.
+ *
+ * This displays the app name,
+ */
+export default class TitleBar extends React.Component {
+  static propTypes = {
+    /**
+     * The title to render.
+     */
+    children: PropTypes.node,
+  };
+
+  static defaultProps = {
+    children: 'Appsemble',
+  };
+
+  render() {
+    const { children } = this.props;
+
+    return (
+      <Portal element={document.getElementsByClassName('navbar')[0]}>
+        <NavbarBrand className={styles.brand}>
+          <NavbarItem>
+            <SideMenuButton />
+          </NavbarItem>
+          <NavbarItem className="title" component="h2">
+            {children}
+          </NavbarItem>
+        </NavbarBrand>
+      </Portal>
+    );
+  }
+}