+++ packages/editor/components/Toolbar/messages.jsx
@@ -1,0 +1,7 @@
+import { defineMessages } from 'react-intl';
+
+export default defineMessages({
+  iconAlt: 'Appsemble logo',
+  docsLink: 'Documentation',
+  logoutButton: 'Logout',
+});